feat(repo): bind fixed overflow to accepted layout
Validate solver-derived fixed rectangles against writable bounds derived from
accepted page profiles, bind them to exact revision/page/block authority, and
return complete typed blocking diagnostics for every crossed edge.

+++ src/backend/diagnostic/domain/lib.rs
@@ -133,7 +133,7 @@ pub enum Evidence {
     /// Exact physical length represented in canonical micrometres.
     PhysicalLength {
         /// Signed canonical physical length.
-        micrometres: i64,
+        micrometres: i128,
         /// Semantic measurement represented by the length.
         quantity: PhysicalLengthQuantity,
     },
